requirements.txt
Flask
flask-cors
Flask-RESTful
flask-swagger-ui
pyodbc
pika
climage
python-logging-loki
PyYAML
SQLAlchemy